requirements.txt
google-auth
opentelemetry-exporter-gcp-trace
google-cloud-discoveryengine
google-cloud-storage 
fastapi 
firecrawl-py 
langchain 
langchain-community 
langchain-openai 
langgraph 
langsmith
opentelemetry-instrumentation-fastapi 
pandas 
psycopg2-binary 
pydantic 
pypdf 
python-dotenv 
redis 
requests 
sqlalchemy 
streamlit 
uvicorn 
yt-dlp
google-cloud-videointelligence
langchain_google_vertexai
langchain_google_community
langchain-google-genai
google-cloud-logging